// src/common/enums/response-message.enum.ts
export enum ResponseMessage {
  SUCCESS = 'Success',
  CREATED = 'Created successfully',
  UPDATED = 'Updated successfully',
  DELETED = 'Deleted successfully',
  BAD_REQUEST = 'Bad request',
  UNAUTHORIZED = 'Unauthorized',
  FORBIDDEN = 'Forbidden',
  NOT_FOUND = 'Not found',
  INTERNAL_ERROR = 'Internal server error',
}